Add files via upload
The code defines a GUI application called "StockAnalyzer" that allows the user to input a ticker symbol for a stock and then provides the user with the option to view basic information, historical data, risk and return, and dividend yield for that stock. The data is fetched from Yahoo Finance API using the "yfinance" library. The GUI is implemented using PyQt5 library. If the user enters an invalid ticker symbol, the program will handle it gracefully and display a message accordingly.